+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 
 * [HexaPDF::Document#wrap] to better handle subtype mappings in case of unknown
   type information
+* [HexaPDF::DictionaryFields::DictionaryConverter] to not allow conversion to a
+  [HexaPDF::Stream] subclass from objects without stream data
 * Import of JPEG images with YCCK color encoding
 * Export of images without `/FlateDecode` filter or `/DecodeParms` to PNG files
 * Mistyped name of field type for field `/Popup` of